Add quotes around go version passed into generated YAML for the go plugin
This was being parsed a numerical YAML value, so `1.20` would be parsed
as `1.2`. Surrounding the version number with quotes solves the
immediate problem. However, it could appear in other places and I did
not check exhaustively. A more reliable approach would be to generate
the pipeline as JSON, parse it as JSON and convert it to YAML.